requirements.txt
torch>=2.1.0
torchvision>=0.16.0
pytorch-lightning>=2.1.0
timm>=0.9.0
gradio>=4.0.0
Pillow>=10.0.0
loguru>=0.7.0
matplotlib>=3.8.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.4.0
tensorboard>=2.14.0
tqdm>=4.66.0
opencv-python>=4.5.0
seaborn==0.13.2